backend/requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
brotlipy==0.7.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
conda==23.1.0
conda-package-handling @ file:///C:/b/abs_fcga8w0uem/croot/conda-package-handling_1672865024290/work
conda_package_streaming @ file:///C:/b/abs_0e5n5hdal3/croot/conda-package-streaming_1670508162902/work
coreapi==2.3.3
coreschema==0.0.4
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
distlib==0.3.6
Django==4.2.4
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
django-snowflake==3.2a2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
filelock==3.9.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
menuinst @ file:///C:/ci/menuinst_1631733438520/work
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==23.1
pipenv==2023.2.4
platformdirs==3.0.0
pluggy @ file:///C:/ci/pluggy_1648024580010/work
psycopg2-binary==2.9.7
pycosat @ file:///C:/b/abs_4b1rrw8pn9/croot/pycosat_1666807711599/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-decouple==3.8
python-nmap==0.7.1
pytz==2022.7.1
pywin32==305.1
PyYAML==6.0.1
referencing==0.30.2
requests @ file:///C:/b/abs_80_edxtvhu/croot/requests_1678709752318/work
rpds-py==0.10.0
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_aarblxbilo/croot/ruamel.yaml.clib_1666302270884/work
simplejson==3.19.1
snowflake-connector-python==3.0.0
sqlparse==0.4.4
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
tqdm @ file:///C:/b/abs_0axbz66qik/croots/recipe/tqdm_1664392691071/work
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
virtualenv==20.19.0
virtualenv-clone==0.5.7
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
zstandard==0.19.0
